Refactor: 로그 삭제
